Move Request to its own file
